Refactoring AvailableNetworks component into src/components/Publish
