+++ src/components/Publish/Title/index.tsx
@@ -5,7 +5,7 @@ import styles from './index.module.css'
 import content from '../../../../content/publish/index.json'
 import { useWeb3 } from '@context/Web3'
 import Info from '@images/info.svg'
-import AvailableNetworks from '@shared/AvailableNetworks'
+import AvailableNetworks from 'src/components/Publish/AvailableNetworks'
 
 export default function Title({
   networkId
